% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binProb.R
\name{binProb}
\alias{binProb}
\title{Binomial CDF calculator}
\usage{
binProb(n, p, min, max)
}
\arguments{
\item{n}{numeric: number of bernoulli trials}

\item{p}{numeric: probability p for each bernoulli trial}

\item{min}{numeric: minimum number of successes}

\item{max}{numeric: maximum number of successes}
}
\value{
numeric: probability that min <= X <= max, given X~Bin(n,p).
}
\description{
Calculates the probability that min <= X <= max, given X~Bin(n,p).
}
\details{
things to say
}
\examples{
# Suppose a student has a 30\% chance they are late to class each day. What is the probability that
# they will be late 1 to 3 days in any given week?
binProb(7, 0.3, 1, 3)
}
